[Feat/#183] 분철팟 검색 기능 구현 (#185)
* [feat/#183] PartySearchScreen 추가

* [feat/#183] Contracts 추가

* [refactor/#183] PotiHeaderPage 수정

* [refactor/#183] 검색 화면 UI 구현

* [feat/#183] 검색 화면 Navigation 연결

* [feat/#183] onSearchKeywordChange 연결

* [chore/#183] 필드 삭제

* [feat/#183] 검색 결과 DTO 추가

* [feat/#183] 검색 결과 DTO 추가

* [feat/#183] 검색 결과 도메인 추가

* [feat/#183] SearchRepository 추가

* [feat/#183] SearchService 추가

* [feat/#183] SearchRemoteDataSource 추가

* [feat/#183] SearchPartyUseCase 추가

* [feat/#183] hasNext 필드 임시 주석 처리

* [feat/#183] 검색 API 연결

* [feat/#183] hasNext 필드 추가

* [feat/#183] 상품 카테고리 분철팟 연결

* [feat/#183] sort 필드 삭제

* [feat/#183] 빈 결과 추가

* [feat/#183] 다음 페이지 로드 시 무한 재시도 제한

* [test/#183] 테스트코드 작성

* [feat/#183] NextPageLoadState 추가

* [chore/#183] 불필요한 코드 삭제

* [chore/#183] Int to Long

* [feat/#183] 내비게이션 UiIntent & UiEffect 사용으로 변경

* [chore/#183] UiMockData 제거

* [feat/#183] 다음 페이지 요청 전에 Loading 상태를 설정

* [feat/#183] 패딩 수정

+++ app/src/main/java/com/poti/android/presentation/party/home/component/GoodsLargeCard.kt
@@ -39,7 +39,7 @@ fun GoodsLargeCard(
     artist: String,
     artistId: Long,
     title: String,
-    partyCount: Int,
+    partyCount: Long,
     tag: String?,
     onClick: (Long, String) -> Unit,
     modifier: Modifier = Modifier,
